pyproject.toml
[tool.poetry]
name = "indicator-of-compromise-enrichment"
version = "0.1.0"
description = ""
authors = ["Furkan Öneryıldız <furkan.oneryildiz@gmail.com>"]
readme = "README.md"
packages = [{include = "indicator_of_compromise_enrichment"}]

[tool.poetry.dependencies]
python = "3.11.4"
fastapi = "^0.100.1"
sqlalchemy = "1.4.42"
databases = {extras = ["postgresql"], version = "^0.7.0"}
uvicorn = {extras = ["standard"], version = "^0.23.0"}
psycopg2-binary = "^2.9.6"
alembic = "^1.11.1"
python-dotenv = "^1.0.0"
celery = "^5.3.1"
redis = "^4.6.0"
gunicorn = "^20.1.0"
requests = "^2.31.0"
geopy = "^2.3.0"
python-whois = "^0.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
